refactor(frontend): the panel's three overlays out of the panel
`panel-dialogs.ts` takes the completion-details dialog, the move-date dialog and
the destructive-action confirmation: their state transitions, their builders,
and the `makeDialog` shell the first two share. `makeDialog` takes no host — it
touches no panel state.

The confirmation's twelve-line teardown — its body-level scrim and both document
keydown listeners — existed three times: opening a confirmation, closing one,
and the panel unmounting mid-dialog. It is now `teardownOverlay`, and
`disconnectedCallback` (which stays in the panel, where unmount belongs) calls
it. `renderConfirmDeleteDialog` deliberately keeps its hand-built scrim on
`document.body` rather than moving onto `makeDialog`: its comment explains why
its Delete has to resolve `variant` against Home Assistant's own theme.

`_openCompletionEdit`, `_openConfirmDialog` and `_openMoveCompletion` leave
`PanelHost` and become module functions their callers import; the dialog state
they act on (`_completion`, `_moveCompletion`, `_confirmDelete`, the scrim and
the two keydown handlers), plus `_liveHassEls`, `_syncDrawerModality` and
`_sensorLive`, join it. `_makeForm` stays on the panel: it is the one `ha-form`
constructor four regions build through, and its whole contract is registering
each form in the panel's own `_liveHassEls`.

Verified byte-identical: the built bundle's template literals are unchanged, and
all three overlays — driven from the real UI, including every dialog form's
schema and data and the scrim's own markup — diff clean against the same
interactions before the move.
